# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterBody" }
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /leads/{id}/activities:
    get:
      operationId: listLeadActivities
      tags: [activities]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeadActivity" }
    post:
      operationId: createLeadActivity
      tags: [activities]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateActivityBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeadActivity" }
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingDetail" }
    patch:
      operationId: updateBooking
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBookingBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
    delete:
      operationId: deleteBooking
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessMessage" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, agent] }
        createdAt: { type: string }
      required: [id, name, email, role, createdAt]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [token, user]
    RegisterBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [admin, agent] }
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
...
    LeadActivity:
      type: object
      properties:
        id: { type: integer }
        leadId: { type: integer }
        userId: { type: integer, nullable: true }
        userName: { type: string, nullable: true }
        type: { type: string, enum: [call, note, whatsapp, email, status_change] }
        description: { type: string }
        createdAt: { type: string }
      required: [id, leadId, type, description, createdAt]
    CreateActivityBody:
      type: object
      properties:
        type: { type: string, enum: [call, note, whatsapp, email] }
        description: { type: string, minLength: 1 }
      required: [type, description]
    Booking:
      type: object
      properties:
        id: { type: integer }
        leadId: { type: integer }
        leadName: { type: string, nullable: true }
        packageName: { type: string }
        destination: { type: string }
        tripType: { type: string, enum: [Domestic, International] }
        startDate: { type: string }
        endDate: { type: string }
        adults: { type: integer }
        children: { type: integer }
        infants: { type: integer }
        totalPersons: { type: integer }
        travelMode: { type: string, enum: [Self, Volvo, Train, Flight] }
        packageServiceType: { type: string, enum: ["Hotel Only", "Hotel + Cab", "Complete Package"] }
        hotelType: { type: string, enum: ["3*", "4*", "5*"], nullable: true }
        mealPlan: { type: string, enum: [CP, MAP, AP], nullable: true }
        costPrice: { type: number }
        salePrice: { type: number }
        profit: { type: number }
        paid: { type: number }
        balance: { type: number }
        createdAt: { type: string }
      required: [id, leadId, packageName, destination, tripType, startDate, endDate, adults, children, infants, totalPersons, travelMode, packageServiceType, costPrice, salePrice, profit, paid, balance, createdAt]
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            payments:
              type: array
              items: { $ref: "#/components/schemas/Payment" }
          required: [payments]
    CreateBookingBody:
      type: object
      properties:
        leadId: { type: integer }
        packageName: { type: string, minLength: 1 }
        destination: { type: string, minLength: 1 }
        tripType: { type: string, enum: [Domestic, International] }
        startDate: { type: string }
        endDate: { type: string }
        adults: { type: integer, minimum: 1 }
        children: { type: integer, minimum: 0 }
        infants: { type: integer, minimum: 0 }
        travelMode: { type: string, enum: [Self, Volvo, Train, Flight] }
        packageServiceType: { type: string, enum: ["Hotel Only", "Hotel + Cab", "Complete Package"] }
        hotelType: { type: string, enum: ["3*", "4*", "5*"], nullable: true }
        mealPlan: { type: string, enum: [CP, MAP, AP], nullable: true }
        costPrice: { type: number, minimum: 0 }
        salePrice: { type: number, minimum: 0 }
      required: [leadId, packageName, destination, tripType, startDate, endDate, adults, children, infants, travelMode, packageServiceType, costPrice, salePrice]
    UpdateBookingBody:
      type: object
      properties:
        packageName: { type: string }
        destination: { type: string }
        tripType: { type: string, enum: [Domestic, International] }
        startDate: { type: string }
        endDate: { type: string }
        adults: { type: integer, minimum: 1 }
        children: { type: integer, minimum: 0 }
        infants: { type: integer, minimum: 0 }
        travelMode: { type: string, enum: [Self, Volvo, Train, Flight] }
        packageServiceType: { type: string, enum: ["Hotel Only", "Hotel + Cab", "Complete Package"] }
        hotelType: { type: string, enum: ["3*", "4*", "5*"], nullable: true }
        mealPlan: { type: string, enum: [CP, MAP, AP], nullable: true }
        costPrice: { type: number, minimum: 0 }
        salePrice: { type: number, minimum: 0 }
    Payment:
      type: object
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        amount: { type: number }
        date: { type: string }
        method: { type: string, enum: [Cash, UPI, "Bank Transfer", Card, Cheque] }
        type: { type: string, enum: [Advance, Partial, Full] }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, bookingId, amount, date, method, type, createdAt]
...