# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityBodyType } from "./createActivityBodyType";

export interface CreateActivityBody {
  type: CreateActivityBodyType;
  /** @minLength 1 */
  description: string;
}